+++ changelog.md
@@ -1,3 +1,18 @@
+## 7.0.0-pre.1 / 2017-07-06
+
+- Add sheets to the registry when it is a new instance
+
+## 7.0.0-pre / 2017-06-28
+
+- Theming - `ThemeProvider` and `withTheme`
+- Update to JSS 8
+- Class generator counter will reset on every request for dynamic SSR.
+
+### Breaking changes
+
+- Removed `createInjectSheet`. Use JssProvider to pass a `jss` instead instead.
+- Renamed `SheetsRegistryProvider` to `JssProvider`
+
 ## 6.1.1 / 2017-04-14
 
 - Use prop-types package
@@ -83,4 +98,3 @@
 - Added a `create()` function to create a new `injectSheet` function which takes a `jss` instance.
 - Updated readme.
 - Added lint-staged.
-